requirements_win32.txt
alembic==1.4.2
aniso8601==8.0.0
apispec==1.3.3
APScheduler==3.6.3
argcomplete==1.11.1
astroid==2.3.3
attrs==19.3.0
azure-core==1.6.0
azure-cosmos==4.0.0
Babel==2.8.0
backcall==0.2.0
bleach==3.1.5
boto3==1.16.43
botocore==1.19.43
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
comtypes==1.1.7
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dnspython==1.16.0
email-validator==1.1.1
entrypoints==0.3
Flask==1.1.2
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
graphviz==0.14
idna==2.10
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pendulum==1.4.4
pickleshare==0.7.5
prison==0.1.3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
PyMySQL==0.9.3
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2020.1
pytzdata==2019.3
pywin32==300
pywinauto==0.6.7
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
wfastcgi==3.0.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
WTForms==2.3.1
zipp==3.1.0
zope.deprecation==4.4.0


